tests/requirements.txt
pyautogui==0.9.41
coloredlogs==10.0
python-dateutil==2.8.0
opencv-python==4.0.0.21
pytesseract==0.2.6
numpy==1.16.1
image==1.5.27
pyperclip==1.7.0
packaging==19.0
gitpython==2.1.11
pynput==1.4.2
mozinstall==2.0.0
mozdownload==1.26
mozversion==2.1.0
mozinfo==1.0.0
mozlog==4.2.0
mozrunner==7.4.0
mss==4.0.1
more-itertools==5.0.0
pytest==5.0.1
pygithub==1.43.8
bugzilla==1.0.0
psutil==5.6.3
funcy==1.11
flaky==3.6.1
selenium==3.141.0
python-dotenv==0.10.3
pypom==2.2.0
